% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{degByBIC}
\alias{degByBIC}
\title{degByBIC}
\usage{
degByBIC(dat, maxorder = 5, weights = NULL, mindeg = 0)
}
\arguments{
\item{dat}{list containing the data for predictors \code{dat$x} and descriptors \code{dat$y}}

\item{maxorder}{integer, maximal order of polynomial model}

\item{weights}{vector of weights for the data in \code{dat}}

\item{mindeg}{integer,  minimal order of polynomial model}
}
\value{
integer, recommended degree for polynomial model
}
\description{
internal function, determines the polynomial model order \code{<=maxorder} for the data \code{dat} which minimizes the BIC information criterion
}
