% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{euclw}
\alias{euclw}
\title{euclw}
\usage{
euclw(x, normalize = T, sto = T)
}
\arguments{
\item{x}{matrix, data.frame}

\item{normalize}{boolean, \code{TRUE} for normalized weights.}

\item{sto}{boolean, if \code{TRUE} ignore higher principal component coordinates.}
}
\value{
matrix
}
\description{
Takes a matrix/data.frame \code{x} with each coloumn having the scores corresponding to  a principal component.
Computes weights as needed in 1d regression models,
i.e. returns for each of the 1d-projections on the PC the Euclidean distances from the n-dimensional point in the n-space.
}
