% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{ftweights}
\alias{ftweights}
\title{ftweights}
\usage{
ftweights(w1)
}
\arguments{
\item{w1}{numeric vector}
}
\value{
numeric vector
}
\description{
Internal function (for \code{pred.patch}, \code{ored.solution}). Replaces infinite numbers (weights) by large finite values. (clipping)
}
