% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{getroots2}
\alias{getroots2}
\title{getroots2}
\usage{
getroots2(dat, degree, target, limit = 40, weights = NULL, retlimit = FALSE)
}
\arguments{
\item{dat}{matrix}

\item{degree}{integer}

\item{target}{numeric vector}

\item{limit}{integer}

\item{weights}{numeric vector}

\item{retlimit}{boolean}
}
\value{

}
\description{
getroots2
}
