% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{mknorm}
\alias{mknorm}
\title{mknorm}
\usage{
mknorm(x, means = NULL, sds = NULL)
}
\arguments{
\item{x}{input vector}

\item{means}{vector of means}

\item{sds}{vector of standard deviations}
}
\value{
standardized vector
}
\description{
internal function, standardizes a vector \code{x} by vector of means \code{means} and vector of standard deviations \code{sds}.
}
