% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{mkreg}
\alias{mkreg}
\title{mkreg}
\usage{
mkreg(x, means, sds)
}
\arguments{
\item{x}{input vector (standardized)}

\item{means}{vector of means}

\item{sds}{vector of standard deviations}
}
\value{
back-transformed vector
}
\description{
internal function, inverse of \code{mknorm}. Back-transformation of a standardized vector.
}
