% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{mvdistance}
\alias{mvdistance}
\title{mvdistance}
\usage{
mvdistance(xs, y, euclid = F)
}
\arguments{
\item{xs}{matrix, each row representing a vector}

\item{y}{vector}

\item{euclid}{\code{TRUE} for Euclidean distance, otherwise MAnhattan distance}
}
\value{
vector of the distances
}
\description{
internal function, computes the Euclidean (\code{euclid=T}) or Manhattan distances (\code{euclid=F}) between vector \code{y} and the rows of matrix \code{xs}.
}
