% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt-package.R
\docType{package}
\name{mvTargetOpt-package}
\alias{mvTargetOpt}
\alias{mvTargetOpt-package}
\title{mvTargetOpt: Multivariate multi-target optimization}
\description{
Multi-objective optimization algorithm.
    A new approach, using ideas from PCA dimension reduction and PLS regression.
}
\keyword{internal}
