% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{nclose2mean}
\alias{nclose2mean}
\title{nclose2mean}
\usage{
nclose2mean(datx, center, n = 1)
}
\arguments{
\item{datx}{matrix}

\item{center}{numeric vector}

\item{n}{integer}
}
\value{
matrix
}
\description{
Internal function. Returns a subset of size \code{n} of the points in \code{datx} which are closest to \code{center}.
}
