% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R, R/mvTargetOptbeforemessingsto.R
\name{newexp}
\alias{newexp}
\title{Title}
\usage{
newexp(n = 25, xpos, foo)

newexp(n = 25, xpos, foo)
}
\arguments{
\item{foo}{}

\item{foo}{}
}
\description{
a helper function for autosolve

a helper function for autosolve
}
\examples{
#not to be used
#not to be used
}
