% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{newexp2}
\alias{newexp2}
\title{newexp2}
\usage{
newexp2(n = 25, xpos, foo, sd = 0.001)
}
\arguments{
\item{n}{integer, number of repeated measurements}

\item{xpos}{numeric matrix, coordinates at which the measurement takes place}

\item{foo}{function, model for the real relationship}

\item{sd}{numeric, standard deviation}
}
\value{
a matrix containing the samples
}
\description{
Creates new samples from model function. Used for simulations in the function autosolve as a helper function.
}
\examples{
#not to be used
}
