% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{PIhcheck}
\alias{PIhcheck}
\title{PIhcheck}
\usage{
PIhcheck(model, alph = 0.05)
}
\arguments{
\item{model}{linear model}

\item{alph}{numeric, confidence level}
}
\value{
vector, upper and lower prediction intervall
}
\description{
internal function, determinas a prediction interval for given linear model \code{model} with confidence level \code{alpha}. currently unused.
}
