% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{plotexp}
\alias{plotexp}
\title{plotexp}
\usage{
plotexp(
  dat,
  target,
  tgerr,
  prediction = NULL,
  model = NULL,
  limit0 = FALSE,
  reality = NULL,
  xlab = "Process parameter (p)",
  ylab = "Descriptor (d)"
)
}
\arguments{
\item{dat}{data.frame, with dat$x matrix of proces parameter data}

\item{target}{numeric vector}

\item{tgerr}{numeric vector}

\item{prediction}{numeric}

\item{model}{function}

\item{limit0}{boolean}

\item{reality}{function}
}
\value{
plot
}
\description{
Function which plots...
}
