% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{ptchoose}
\alias{ptchoose}
\title{ptchoose}
\usage{
ptchoose(
  ptchoice = ptchoice,
  dat1d,
  tgmean_norm,
  maxarea,
  xmeans,
  xsds,
  pls1,
  jjj
)
}
\arguments{
\item{ptchoice}{numeric, 1 to continue search outside the explored parameters}

\item{dat1d}{data.frame, predictor data must be in \code{dat1d$x}}

\item{tgmean_norm}{numeric vector}

\item{maxarea}{NULL or matrix}

\item{xmeans}{numeric vector}

\item{xsds}{numeric vector}

\item{pls1}{result of the PLS1 function}

\item{jjj}{integer, counts how often this function has been called}
}
\value{
matrix
}
\description{
Internal function. Continues search in unexplored space, when no solution can be found.
}
