% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{sumthelowers}
\alias{sumthelowers}
\title{sumthelowers}
\usage{
sumthelowers(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
internal helper function which is used by \code{euclw} .
}
