% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{sumtheothers}
\alias{sumtheothers}
\title{sumtheothers}
\usage{
sumtheothers(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
internal helper function which is used by \code{euclw} .
}
