% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{uniqP}
\alias{uniqP}
\title{uniqP}
\usage{
uniqP(newx, xeps, datx)
}
\arguments{
\item{newx}{matrix}

\item{xeps}{numeric}

\item{datx}{matrix}
}
\value{
matrix
}
\description{
Internal function.
Returns a subset of the  unique points of \code{newx} whgich are not
(up to an epsilon \code{xeps}) identical to the points in \code{datx}.
}
