% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usejson.R
\name{usejson}
\alias{usejson}
\title{usejson}
\usage{
usejson()
}
\value{
json file with suggested parameters and flag (is parameter set a prediction or simply a suggestion how to improve the model)
}
\description{
apply algorithm to some data from json files
}
\examples{
setwd("C:/Users/maendle/OneDrive/workingr/tgopt")

}
