% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{autosolve}
\alias{autosolve}
\title{Autosolve / PLSolve.auto}
\usage{
autosolve(
  startx,
  tgmean,
  tgerr,
  reps = 25,
  maxit = 10,
  reality = foo,
  xeps = 0.01,
  pplot = TRUE,
  pcnr = c(1, 2),
  additive = FALSE,
  maxarea = NULL,
  useweights = TRUE,
  mknormweights = F,
  gr2retlimit = T,
  mindeg = 0,
  sequential = F,
  tgpcawg = 1,
  yweights = F,
  datlim = NULL,
  knearest = NULL,
  tgdim = 1,
  ylast = NULL,
  sto = T,
  mod.sd = NULL,
  ...
)
}
\arguments{
\item{startx}{start values}

\item{tgmean}{target value vector}

\item{tgerr}{vector of acceptable error range}

\item{reps}{number of repeated measurements}

\item{maxit}{maximum number of iterations}

\item{reality}{real model}

\item{xeps}{smallest unit?}

\item{pplot}{do plots?}

\item{pcnr}{which principal directions will be considered}

\item{additive}{use soem additive approach (deprecated)}
}
\value{

}
\description{
Autosolve / PLSolve.auto
}
\examples{
# Erzeuge 3x3 Modell für Simulation
startx <- mydata
tfoo <- function(x) {
  x1 <- x[,1]
  x2 <- x[,2]
  # x3 <- x[,3]
  return( data.frame( y1=0.8*x1 - 1.2*x2,
                      y2=0.8*x1 - 1.2*abs(x2)^0.25#,
                     #  y3=0.8*x1 - 0.6*abs(x2)^0.25 + x3
 ) ) }
tgmean <- tfoo(cbind(1.35,1.4))#,1.5))#tfoo(cbind(0.35,0.4,0.5))
tgerr <- c(0.2,0.2)#,0.5)
#reps <- 5
#maxit <-10
#reality <- tfoo
#xeps <- 0.01
#pplot <- TRUE
#additive=TRUE
tmp<-autosolve(startx,tgmean,tgerr*0.125,reps=7,maxit=10,tfoo, xeps=0.01, F, pcnr=c(1,2), additive=T)
#nwexample
dstgmean <- tfoo(cbind(1.35,1.4))#,1.5))#tfoo(cbind(0.35,0.4,0.5))
tgerr <- c(0.2,0.2)#,0.5)
reps <- 5
maxit <-10
reality <- tfoo
xeps <- 0.01
pplot <- TRUE
additive=TRUE
startx <- expand.grid(x1=c(-1,3),x2=c(-1,3))
set.seed(123)
autosolve(startx,tgmean,tgerr*0.125,reps=7,maxit=10,tfoo, xeps, F, pcnr=c(1,2), additive=F)
# ? zurückgeben mit Welcher WK man im Ziel landet
}
