% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R
\name{newexp}
\alias{newexp}
\title{newexp}
\usage{
newexp(n = 25, xpos, foo, sd = 0.001)
}
\arguments{
\item{n}{number of repeated measurements}

\item{xpos}{coordinates at which the measurement takes place}

\item{foo}{model for the real relationship}
}
\value{
a matrix containing the samples
}
\description{
Creates new samples from model function. Used for simulations in the function autosolve as a helper function.
}
\examples{
#not to be used
}
