% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvTargetOpt.R, R/mvTargetOptbeforemessingsto.R
\name{tgpca}
\alias{tgpca}
\title{PCA}
\usage{
tgpca(dat, tgmean = NULL, tgerr = NULL, wg = 0.901, wfun,
  mknormweights, yweights = F, ylast = NULL)

tgpca(dat, tgmean = NULL, tgerr = NULL, wg = 0.901, wfun,
  mknormweights, yweights = F, ylast = NULL)
}
\arguments{
\item{tgmean}{optional}

\item{tgerr}{optional}

\item{wg}{weight of the target value. If wg equals 1 or 2 then the pca is performed with the target value as center}

\item{tgmean}{optional}

\item{tgerr}{optional}

\item{wg}{weight of the target value. If wg equals 1 or 2 then the pca is performed with the target value as center}
}
\value{
returns the results of the pca and some extra stuff

returns the results of the pca and some extra stuff
}
\description{
used for dimension reduction of the target space
in order to choose a certain direction through the target point for the projections wg has to be set to 1, then the target point is chosen as center for the pca.
If wg between 0 and 1 then pseudo observations at the target point are created such that a ratio of wg of the observations are pseudo observations.
uses prcomp on standardized data and pseudo data.

used for dimension reduction of the target space
in order to choose a certain direction through the target point for the projections wg has to be set to 1, then the target point is chosen as center for the pca.
If wg between 0 and 1 then pseudo observations at the target point are created such that a ratio of wg of the observations are pseudo observations.
uses prcomp on standardized data and pseudo data.
}
\examples{
tgpca(matrix(rnorm(20),ncol=2))
tgpca(matrix(rnorm(20),ncol=2))
}
